+++ cmd/conductor/main.go
@@ -43,6 +43,8 @@ func init() {
 	rootCmd.AddCommand(statusCmd)
 	rootCmd.AddCommand(tuiCmd)
 	rootCmd.AddCommand(versionCmd)
+	rootCmd.AddCommand(updateCmd)
+	rootCmd.AddCommand(migrateCmd)
 }
 
 var versionCmd = &cobra.Command{